Remove FlattenExceptionTrait
