+++ composer.json
@@ -24,6 +24,7 @@
         "symfony/config": "^4.2|^5.0",
         "symfony/dependency-injection": "^4.2|^5.0",
         "symfony/doctrine-bridge": "^4.2|^5.0",
+        "symfony/error-handler": "^4.4|^5.0",
         "symfony/event-dispatcher": "^4.2|^5.0",
         "symfony/expression-language": "^4.2|^5.0",
         "symfony/finder": "^4.2|^5.0",
