+++ src/Exception/FlattenException.php
@@ -2,17 +2,60 @@
 
 namespace EasyCorp\Bundle\EasyAdminBundle\Exception;
 
-use Symfony\Component\Debug\Exception\FlattenException as LegacyBaseFlattenException;
 use Symfony\Component\ErrorHandler\Exception\FlattenException as BaseFlattenException;
 
-if (class_exists(BaseFlattenException::class)) {
-    class FlattenException extends BaseFlattenException
+/**
+ * @author Maxime Steinhausser <maxime.steinhausser@gmail.com>
+ */
+class FlattenException extends BaseFlattenException
+{
+    /** @var ExceptionContext */
+    private $context;
+
+    /**
+     * @param \Exception $exception
+     * @param int        $statusCode
+     * @param array      $headers
+     *
+     * @return FlattenException
+     *
+     * @throws \RuntimeException
+     */
+    public static function create(\Exception $exception, $statusCode = null, array $headers = []): parent
     {
-        use FlattenExceptionTrait;
+        if (!$exception instanceof BaseException) {
+            throw new \RuntimeException(sprintf('You should only try to create an instance of "%s" with a "EasyCorp\Bundle\EasyAdminBundle\Exception\BaseException" instance, or subclass. "%s" given.', __CLASS__, \get_class($exception)));
+        }
+
+        /** @var FlattenException $e */
+        $e = parent::create($exception, $statusCode, $headers);
+        $e->context = $exception->getContext();
+
+        return $e;
     }
-} else {
-    class FlattenException extends LegacyBaseFlattenException
+
+    public function getPublicMessage()
     {
-        use FlattenExceptionTrait;
+        return $this->context->getPublicMessage();
+    }
+
+    public function getDebugMessage()
+    {
+        return $this->context->getDebugMessage();
+    }
+
+    public function getParameters()
+    {
+        return $this->context->getParameters();
+    }
+
+    public function getTranslationParameters()
+    {
+        return $this->context->getTranslationParameters();
+    }
+
+    public function getStatusCode(): int
+    {
+        return $this->context->getStatusCode();
     }
 }
